+++ tests/testthat/test_matchMz.R
@@ -62,6 +62,8 @@ test_that("matchMz,Mass2MzParam works", {
   expect_equal(res@matches$query_idx, c(1, 2, 2, 3))
   expect_equal(res@matches$target_idx, c(1, 2, 3, 1))
   expect_equal(res@matches$score, c(0, 0, 0, 1e-6))
+  expect_equal(res@matches$ppm_error,
+               c(0, 0, 0, 1 / mass2mz(204.089878, "[M+Na]+")))
 
   par <- Mass2MzParam(adducts = adducts, tolerance = 0, ppm = 0)
   res <- matchMz(x, cmpds, par)
@@ -70,6 +72,7 @@ test_that("matchMz,Mass2MzParam works", {
   expect_equal(res@matches$query_idx, c(1, 2, 2))
   expect_equal(res@matches$target_idx, c(1, 2, 3))
   expect_equal(res@matches$score, c(0, 0, 0))
+  expect_equal(res@matches$ppm_error, c(0, 0, 0))
 
   ## no matches
   adducts <- c("[M+Li]+", "[M+K]+")
@@ -86,7 +89,7 @@ test_that("matchMz,Mass2MzParam works", {
   x <- data.frame(
     mz = c(mass2mz(204.089878, adducts["a", ]),
            mass2mz(131.094629, adducts["a", ]),
-           mass2mz(204.089878, adducts["b", ]) + 1e-6)
+           mass2mz(204.089878, adducts["b", ]) - 2e-6)
   )
 
   par <- Mass2MzParam(adducts = adducts, tolerance = 0, ppm = 20)
@@ -95,7 +98,9 @@ test_that("matchMz,Mass2MzParam works", {
   expect_equal(target(res), cmpds)
   expect_equal(res@matches$query_idx, c(1, 2, 2, 3))
   expect_equal(res@matches$target_idx, c(1, 2, 3, 1))
-  expect_equal(res@matches$score, c(0, 0, 0, 1e-6))
+  expect_equal(res@matches$score, c(0, 0, 0, 2e-6))
+  expect_equal(res@matches$ppm_error,
+               c(0, 0, 0, -2 / mass2mz(204.089878, adducts["b", ])))
 
   par <- Mass2MzParam(adducts = adducts, tolerance = 0, ppm = 0)
   res <- matchMz(x, cmpds, par)
@@ -104,6 +109,7 @@ test_that("matchMz,Mass2MzParam works", {
   expect_equal(res@matches$query_idx, c(1, 2, 2))
   expect_equal(res@matches$target_idx, c(1, 2, 3))
   expect_equal(res@matches$score, c(0, 0, 0))
+  expect_equal(res@matches$ppm_error, c(0, 0, 0))
 
   expect_error(matchMz(x, cmpds, par, massColname = "other"), "other")
 
@@ -114,6 +120,8 @@ test_that("matchMz,Mass2MzParam works", {
   expect_equal(res@matches$query_idx, c(1, 2, 2))
   expect_equal(res@matches$target_idx, c(1, 2, 3))
   expect_equal(res@matches$score, c(0, 0, 0))
+  expect_equal(res@matches$ppm_error, c(0, 0, 0))
+  
 
   expect_error(matchMz(x$mz, cmpds, par, massColname = "other"), "other")
 
@@ -124,6 +132,8 @@ test_that("matchMz,Mass2MzParam works", {
   expect_equal(res@matches$query_idx, c(1, 2, 2))
   expect_equal(res@matches$target_idx, c(1, 2, 3))
   expect_equal(res@matches$score, c(0, 0, 0))
+  expect_equal(res@matches$ppm_error, c(0, 0, 0))
+  
 
   expect_error(matchMz(x, cmpds$exactmass, par, mzColname = "other"), "other")
 
@@ -182,6 +192,7 @@ test_that("matchMz,Mass2MzRtParam works", {
   expect_equal(res@matches$query_idx, c(1, 2, 2))
   expect_equal(res@matches$target_idx, c(1, 2, 3))
   expect_equal(res@matches$score, c(0, 0, 0))
+  expect_equal(res@matches$ppm_error, c(0, 0, 0))
   expect_equal(res@matches$score_rt, c(0, 0, 0))
 
   par <- Mass2MzRtParam(adducts = adducts, tolerance = 0, ppm = 20,
@@ -192,6 +203,8 @@ test_that("matchMz,Mass2MzRtParam works", {
   expect_equal(res@matches$query_idx, c(1, 2, 2, 3))
   expect_equal(res@matches$target_idx, c(1, 2, 3, 1))
   expect_equal(res@matches$score, c(0, 0, 0, 1e-6))
+  expect_equal(res@matches$ppm_error,
+               c(0, 0, 0, 1 / mass2mz(204.089878, "[M+Na]+")))
   expect_equal(res@matches$score_rt, c(0, 0, 0, 0.1))
 
   par <- Mass2MzRtParam(adducts = adducts, tolerance = 0, ppm = 0,
@@ -202,6 +215,7 @@ test_that("matchMz,Mass2MzRtParam works", {
   expect_equal(res@matches$query_idx, c(1, 2, 2))
   expect_equal(res@matches$target_idx, c(1, 2, 3))
   expect_equal(res@matches$score, c(0, 0, 0))
+  expect_equal(res@matches$ppm_error, c(0, 0, 0))
   expect_equal(res@matches$score_rt, c(0, 0, 0))
 
 
@@ -286,7 +300,8 @@ test_that("matchMz, MzParam works", {
   expect_equal(res@matches$query_idx, c(1, 2))
   expect_equal(res@matches$target_idx, c(5, 7))
   expect_equal(res@matches$score, c(0, 0))
-
+  expect_equal(res@matches$ppm_error, c(0, 0))
+  
   ## no matches
   res <- matchMz(qry + 0.1, trgt, par)
   expect_true(is(res, "Matched"))
@@ -302,6 +317,8 @@ test_that("matchMz, MzParam works", {
   expect_equal(res@matches$query_idx, c(1, 1, 1, 2, 2, 2, 3, 3))
   expect_equal(res@matches$target_idx, c(4, 5, 6, 6, 7, 8, 7, 8))
   expect_equal(res@matches$score, c(10, 0, 10, 10, 0, 10, 9, 1))
+  expect_equal(res@matches$ppm_error, c(10, 0, -10, 10, 0, -10, 9, -1) /
+                 c(140, 150, 160, 160, 170, 180, 170, 180) * 10^6)
 })
 
 test_that("matchMz, MzRtParam works", {
@@ -316,6 +333,7 @@ test_that("matchMz, MzRtParam works", {
   expect_equal(res@matches$query_idx, c(1))
   expect_equal(res@matches$target_idx, c(3))
   expect_equal(res@matches$score, c(0))
+  expect_equal(res@matches$ppm_error, c(0))
   expect_equal(res@matches$score_rt, c(0))
 
   par <- MzRtParam(tolerance = 0.1, ppm = 0, toleranceRt = 0)
@@ -323,6 +341,7 @@ test_that("matchMz, MzRtParam works", {
   expect_equal(res@matches$query_idx, c(1, 2))
   expect_equal(res@matches$target_idx, c(3, 4))
   expect_equal(res@matches$score, c(0, 0.1))
+  expect_equal(res@matches$ppm_error, c(0, 0.1 / 14 * 10^6))
   expect_equal(res@matches$score_rt, c(0, 0))
 
   par <- MzRtParam(tolerance = 0.1, ppm = 0, toleranceRt = 0.2)
@@ -330,6 +349,7 @@ test_that("matchMz, MzRtParam works", {
   expect_equal(res@matches$query_idx, c(1, 2, 3))
   expect_equal(res@matches$target_idx, c(3, 4, 7))
   expect_equal(res@matches$score, c(0, 0.1, 0))
+  expect_equal(res@matches$ppm_error, c(0, 0.1 / 14 * 10^6, 0))
   expect_equal(res@matches$score_rt, c(0, 0, 0.2))
 
   ## no matches
@@ -355,6 +375,7 @@ test_that("matchMz, MzRtParam works", {
   expect_equal(res@matches$query_idx, c(1))
   expect_equal(res@matches$target_idx, c(3))
   expect_equal(res@matches$score, c(0))
+  expect_equal(res@matches$ppm_error, c(0))
   expect_equal(res@matches$score_rt, c(0))
 
   par <- MzRtParam(tolerance = 0.1, ppm = 0, toleranceRt = 0)
@@ -362,6 +383,7 @@ test_that("matchMz, MzRtParam works", {
   expect_equal(res@matches$query_idx, c(1, 2))
   expect_equal(res@matches$target_idx, c(3, 4))
   expect_equal(res@matches$score, c(0, 0.1))
+  expect_equal(res@matches$ppm_error, c(0, 0.1 / 14 * 10^6))
   expect_equal(res@matches$score_rt, c(0, 0))
 
   par <- MzRtParam(tolerance = 0.1, ppm = 0, toleranceRt = 0.2)
@@ -369,6 +391,7 @@ test_that("matchMz, MzRtParam works", {
   expect_equal(res@matches$query_idx, c(1, 2, 3))
   expect_equal(res@matches$target_idx, c(3, 4, 7))
   expect_equal(res@matches$score, c(0, 0.1, 0))
+  expect_equal(res@matches$ppm_error, c(0, 0.1 / 14 * 10^6, 0))
   expect_equal(res@matches$score_rt, c(0, 0, 0.2))
 
   ## no matches